04_AI Agents Healthcare/requirements.txt
bcrypt==4.1.3
build==1.2.1
chroma-hnswlib==0.7.3
chromadb==0.3.26
fqdn==1.5.1
holoviews==1.18.3
httptools==0.6.1
ibm_watson_machine_learning==1.0.360
importlib_resources==6.4.0
isoduration==20.11.0
kubernetes==30.1.0
langchain==0.1.10
langchain-ibm==0.2.1
langchain-openai==0.3.3
langgraph==0.2.69
matplotlib==3.8.4
mmh3==4.1.0
notebook==7.1.2
numexpr==2.10.1
opensearch-py==2.8.0
pipdeptree==2.25.0
pre_commit==4.0.1
pyarrow==16.0.0
PyPika==0.48.9
python-dotenv==1.0.1
sentence-transformers==3.0.1
statsmodels==0.14.1
unstructured==0.14.10
uri-template==1.3.0
uvloop==0.19.0
watchfiles==0.22.0
webcolors==1.13
websockets==12.0
wget==3.2
wheel==0.43.0
